feat: extract visibleViewController in UIApplication extension and use it
